Major Changes on Route Selection and Circular Overlays
Initially, the waiting status of the user is set to "No".

If the user toggles the waiting status to "Yes", the user is now forced to select his destination.

The destination of each user is now being recorded in firebase realtime database.

For the circular overlays, those that have a value of 0 is now invisible. 

The value of the overlays are updating without refreshing the site. The flicker of the overlays mean that there is an ongoing update within the location or waiting status of the users. 

The "Select Destination" below the map also updates the destination of the users in the database.