Circular Overlays Route Alignment
The overlays can now filter out the waiting users that are not on the selected destination. This will enable jeepney drivers to monitor only the waiting users that are in the same destination as they are.

Also, a daily counter of the number of devices who accessed the site in a day was implemented.

The criteria for what is being considered as a waiting user was also changed. Now, a waiting commuter means that they declared that they are waiting (waitingStatus = Yes) and they are online (status = Online).

The issue wherein the number in the circular overlay is not being shown when the width of the screen is large is also fixed.